+++ HISTORY.rst
@@ -3,9 +3,12 @@
 History
 -------
 
-1.4.2 (unreleased)
-++++++++++++++++++
+1.5 (unreleased)
+++++++++++++++++
 
+* Added the ability to do
+  meth:`~django_functest.ShortcutLoginMixin.shortcut_login` with just a user
+  object (no password needed).
 
 1.4.1 (2022-07-12)
 ++++++++++++++++++